Creating Login screen UI
Not all styling is working; for example - inside of the Sign In box, it needs more responsive padding at the top and bottom. The Google login button will be a different type of component once OAuth works. And probably want to ask the Tradecraft guys how they want this page to look at smaller screen widths (assuming I should just center the column instead of offsetting it on the right).